requirements.txt
django==3.0.2
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0
channels==2.4.0
channels_redis==2.4.1
